requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
python-dotenv==1.0.0
pyotp==2.9.0
SmartApi-Python==1.3.5
pydantic==2.5.0
aiofiles==23.2.1
rich==13.7.0
